leave swapfile enabled for Pi Zero 2 W
this also required to enable "Persistent Logging" after the first boot

+++ setup-pi4-viruspilot.sh
@@ -7,9 +7,11 @@ timedatectl set-timezone Europe/Berlin
 apt install libjpeg62-turbo-dev libconfig9 rpi-update dnsmasq git cmake libusb-1.0-0-dev build-essential \
   autoconf libtool i2c-tools libfftw3-dev libncurses-dev python3-serial jq ifplugd iptables -y
 
-# disable swapfile
-systemctl disable dphys-swapfile
-apt purge dphys-swapfile -y
+# disable swapfile for Pi3B and Pi4B
+if ! grep -q grep "Pi Zero 2 W" /sys/firmware/devicetree/base/model; then
+  systemctl disable dphys-swapfile
+  apt purge dphys-swapfile -y
+fi
 
 # cleanup
 apt autoremove -y
